Implemented working version of snapshot loading
Snapshots are really just copies of the data from a particular point in
time and so shouldn't contain any 'live' objects.  This has been
implemented now by loading from the archive all references when a
snapshot is being loaded (except those loaded as part of that
transaction which are just returned).