Fixed a bug in historian and live referencer
Historian when saving was not comparing with a loaded snapshot to see if
the current version had changed but rather a snapshot that included
references to live objects.

Live referencer was getting references from the historian during a save
which it shouldn't do.  They either to be in the transaction or should
be resaved.