Removed historian from SavableObject constructor
This change allows us to move the creation registering to a separate
object that does not need the historian to be active.  This fixes a
problem with, say, module level variables that were instantiated and
straight away needed the historian.

The historian is not captured only when the object is saved with the
historian.  This makes more sense and also removes the awkwardness of
subclasses neededing to pass up the constructor if the object should
support being assigned to a particular historian.