Significant database speed improvement
The MongoDB backend has been partitioned into two collections: data and
history.  Only live objects are kept in the data collection while older
versions and deleted objects are kept in history.  I've added benchmarks
and this results in an over 10x speedup when finding objects when the
dataset is larger.  The scaling should also be much better as now
removed objects and history will not be considered when searching live
objects.

A database migration is required as the schema has changed as follows:

Data collection
---------------

`_id` is now the object id
All historical and deleted records are removed.

History collection
------------------

Contains what data used to contain (including currently live objects).
The _id field is the sref string i.e. <obj_id>#<version>

References collection
---------------------

Has been changed because of the format change of history ids.  This can
simply be dropped as it will be lazily re-created if needed.

Meta collection
---------------

Unchanged.