Fixed additional issues related to delete()
This is a tricky one to get right when using transactions as the object
hasn't been deleted yet in the database.  Now the behaviour will be the
same between in-transaction and outside i.e. a NotFound exception will
be raised if an object that has been deleted is deleted for a second
time.