Added tests for saving metadata during object save
In save_instance_state it is possible to set the metadata on an object
directly.  This is very useful in a lot of cases.  For example, if an
object is being saved and needs certain metadata to be available for
easy searching but that metadata is a derivative of what's in the state
then this is the perfect place to put it because the metadata will
always be up to date wih the current version of the object.