First implementation of object migrations
Objects can now change the way that they store things in the database by
providing a migration class class each time it's changed.  This way,
using a string of migrations, data can always be made compatible with
the current codebase.

I've added tests for several possible scenario but these are by no means
complete.

NOTE: this version does not change the data in the database, rather it
lazily migrates it as needed.  Updating the records in the database with
the new version is the next step.