Fixed a couple of bugs
AsRef in SimpleObject wasn't working correctly when there was more than
one specification for the same attribute (name) in the MRO.  This has
been resolved and the code simplified by making AsRef a factory for an
AttrSpec which is the object that is really used by the code.

The migration tests had a logic error in the way that references work
which has been fixed and additional tests added.