Deprecated get_reference and SnapshotRef
The snapshot reference is now called the snapshot id because this helps
to diffrentiate it from the ObjRef which is a true reference in the
programming sense, in that it can be used directly to retrieve the
instance it is referring to while the snapshot id just identifies the
corresponding object.

Have put deprecation warnings and the old methods/classes will be
dropped in the 0.14.