Fixed mincepy_gui issue 11
This was related to large databases where a query was made involving
sorting.  MongoDB would run out of memory causing an error.  Enabled the
use of disk based storage in such situations when find()ing objects.

Also found an optimisation in the WeakIdObjDict which was causing the
test to take a long time.