Changed operations to use in_transaction()
This context allows the historian to either operate within the current
transaction or create a new one if one does not already exist.  This is
probably the most commonly desired behaviour as if the user calls a
historian function within a transaction they likely want constraint
checks to be deferred till the end of their transaction (rather than at
the end of the historian call e.g. delete()).

This may also offer some speed advantages as there are fewer overall
nested transactions and therefore the cost of creating, checking and
cleaning up each nested transaction is removed.  This may be most
prominently seen on objects that have deeply nested references that each
need to be saved.