Fix for historian find
Historian find() was changed to only take a list of obj ids or obj
types.  This is because it is not always obvious if a passed type is a
collection of, for example, obj ids or is itself an obj id.  We've now
made this work by checking for the latter first and then the former.