Fixed couple of issues related to Savables
The WithFields metaclass had a __getitem__ method which was making
python think it was iterable (which it shouldn't be).  This is a legacy
from python's past so the solution has been to make it a fully fledged
mapping with __iter__ and __len__.  This way it gets the correct
(string) fields.