Optimised decoding
Decoding will now skip trying to further decode primitive,
non-container, types.  This was needlessly causing checks and recursive
functions calls which could add up to a noticable runtime cost in
certain scenarios.

The recursive nature of decode() is still a bit of a concern and future
optimisation may well be worth looking into.