Tidied up historian/depositors
Now, a lot of the functionality related to loading and saving is in the
depositors.  This is a better compartmentisation as before this was
spread between the historian and depositor.  Now the depositor has a
very clear responsibility of acting as a bridge between the archive and
the historian when loading or saving objects.