Migrated metadata to be part of data collection
The MongoDB backend no longer has separate collections for the data
records and metadata.  This is because MongoDB is document based as
doing these kinds of joins is very slow and pointless when there is a
one-to-one mapping as in the case of data record and metadata.  Thus, we
now store the metadata in the 'meta' field of each data record document.

The migration is performed automatically and all indexes found on the
old metadata collection are transferred over.