Fixex bug in meta_update
If a user tried to update the metadata with a set of key: value pairs
that are identical to those already set on the data, then the archive
would raise an exception saying that it couldn't find the object to
update metadata for.  This happened because of checking for the number
of modified documents (which is zero in this case as the user is not
changing the metadata), instead of the number of matched documents.