Moved TupleHelper down proprity list
Previously, mincepy would provide TupleHelper as a plugin (via entry
points) which could mean that it would have a higher priority than a
user-defined type that subclasses tuple.  In such a case, mincepy's own
TupleHelper would be used to perform saving/loading.  This should not be
the case, and the TupleHelper should be there for tuple types or as a
last resort where a user's code might be fine with accepting a tuple
instead of their custom type.

In any case, this has been resolved by making TupleHelper a 'builtin'
type, meaning that it's always savable by the Historian.